multimed_fusion_be/requirements.txt
# Django and Core
Django==5.1.5
djangorestframework==3.15.0
django-cors-headers==4.4.0
django-mongodb-backend==5.1.5

# JWT Authentication
djangorestframework-simplejwt==5.3.2

# Environment Variables
python-dotenv==1.0.1

# PDF Processing
pypdf==5.1.0

# Document Processing
python-docx==1.2.2
pydicom==2.4.4

# Text Embeddings (FastEmbed)
fastembed==0.3.5

# Numerical Computing
numpy==1.26.4

# Database/ORM
pymongo==4.10.1

# Audio Processing (Faster & Lightweight)
faster-whisper==1.0.3

# Image OCR (Fast)
pytesseract==0.3.10

# Pillow for image processing
Pillow==10.2.0

# Gemini AI Integration
google-generativeai==0.3.0

# Utilities
requests==2.32.3

